Fixed default print/expr styles.
Previously, the initial default style was a "poor man's" version
that no one would want.  The actually intended default style was
implemented elsewhere and installed in the parameter via side-effect.
This resulted in some programs getting the wrong default print or
expression style, either due to dependencies or module ordering or
threads.

This commit reimplements the "stylish" printing system using units
to resolve the dependencies so that printing styles and the
printing engine can depend one each other, and the style parameters
can have the right default from all perspectives.